+++ pyinterpolate/processing/select_values.py
@@ -13,7 +13,8 @@ import pandas as pd
 
 from scipy.linalg import fractional_matrix_power
 
-from pyinterpolate.distance.distance import calc_point_to_point_distance, calc_block_to_block_distance
+from pyinterpolate.distance.distance import calc_point_to_point_distance, calc_block_to_block_distance, \
+    calc_angles, calculate_angular_distance
 from pyinterpolate.processing.preprocessing.blocks import Blocks
 from pyinterpolate.processing.transform.transform import get_areal_centroids_from_agg, transform_ps_to_dict
 
@@ -32,8 +33,8 @@ def _rotation_matrix(angle: float) -> np.array:
         The rotation matrix.
     """
     theta = np.radians(angle)
-    e_major_rot = [np.cos(theta), -np.sin(theta)]
-    e_minor_rot = [np.sin(theta), np.cos(theta)]
+    e_major_rot = [np.cos(theta), np.sin(theta)]
+    e_minor_rot = [-np.sin(theta), np.cos(theta)]
     e_matrix = np.array([e_major_rot, e_minor_rot])
     return e_matrix
 
@@ -103,7 +104,7 @@ def select_points_within_ellipse(ellipse_center: np.array,
                 Step size between lags.
 
     theta : float
-            Angle from y axis clockwise (W-E is a 0).
+            Angle from y axis counterclockwise (W-E is a 0).
 
     minor_axis_size : float
                       Fraction of the major axis size.
@@ -169,6 +170,15 @@ def generate_triangles(points: np.ndarray, step_size: float, angle: float, toler
     ----------
     points : numpy array
         The points to find their neighbors.
+
+    step_size : float
+        Lag length.
+
+    angle : float
+        The direction of a variogram.
+
+    tolerance : float
+        The parameter to control ratio of triangle base to its height.
 
     Returns
     -------
@@ -440,6 +450,121 @@ def prepare_pk_known_areas(point_support_dict: Dict,
     return datasets
 
 
+def select_possible_neighbors_angular(possible_neighbors: np.ndarray,
+                                      distances: np.ndarray,
+                                      angle_differences: np.ndarray,
+                                      max_range: float,
+                                      min_number_of_neighbors: int,
+                                      use_all_neighbors_in_range: bool) -> np.ndarray:
+    """
+    Function selects possible neighbors.
+
+    Parameters
+    ----------
+    possible_neighbors : numpy array
+        The array with possible neighbors.
+
+    distances : numpy array
+        The array with distances to each neighbor.
+
+    angle_differences : numpy array
+        The array with the minimal angle between expected distance and other points.
+
+    max_range : float
+        Range within neighbors are affecting the value, it should be lower or the same as the variogram range.
+
+    min_number_of_neighbors : int
+        Number of the n-closest neighbors used for interpolation.
+
+    use_all_neighbors_in_range : bool
+        ``True``: if number of neighbors within the ``neighbors_range`` is greater than the ``number_of_neighbors``
+        then take all of them for modeling.
+
+    Returns
+    -------
+    possible_neighbors : numpy array
+        Sorted neighbors based on a distance and angle from the origin.
+    """
+    angular_tolerance = 1
+    max_tick = 15
+
+    neighbors_dists_and_angles = np.c_[possible_neighbors, distances.T, angle_differences.T]
+
+    # First select only neighbors in a max range
+    sorted_neighbors_and_dists = neighbors_dists_and_angles[neighbors_dists_and_angles[:, -2].argsort()]
+    prepared_data = sorted_neighbors_and_dists[sorted_neighbors_and_dists[:, -2] <= max_range, :]
+
+    # Limit to a specific angle
+    prepared_data_with_angles = prepared_data[prepared_data[:, -1] <= angular_tolerance]
+
+    while (len(prepared_data_with_angles) < min_number_of_neighbors) or (max_tick > angular_tolerance):
+        angular_tolerance = angular_tolerance + 1
+        prepared_data_with_angles = prepared_data[prepared_data[:, -1] <= angular_tolerance]
+
+    if len(prepared_data_with_angles) >= min_number_of_neighbors:
+        if use_all_neighbors_in_range:
+            return prepared_data_with_angles
+        else:
+            return prepared_data_with_angles[:min_number_of_neighbors]
+    else:
+        sorted_with_sorted_angles = sorted_neighbors_and_dists[sorted_neighbors_and_dists[:, -1].argsort()]
+        return sorted_with_sorted_angles[:min_number_of_neighbors]
+
+
+def select_kriging_data_from_direction(unknown_position: Iterable,
+                                       data_array: np.ndarray,
+                                       neighbors_range: float,
+                                       direction: float,
+                                       number_of_neighbors: int = 4,
+                                       use_all_neighbors_in_range: bool = False) -> np.ndarray:
+    """
+    Function selects closest neighbors based on the specific angle and tolerance.
+
+    Parameters
+    ----------
+    unknown_position : Iterable
+        List, tuple or array with x, y coordinates.
+
+    data_array : numpy array
+        Known points.
+
+    neighbors_range : float
+        Range within neighbors are affecting the value, it should be lower or the same as the variogram range.
+
+    direction : float
+        The angle of a directional variogram.
+
+    number_of_neighbors : int, default = 4
+        Number of the n-closest neighbors used for interpolation.
+
+    use_all_neighbors_in_range : bool, default = False
+        ``True``: if number of neighbors within the ``neighbors_range`` is greater than the ``number_of_neighbors`` then
+        take all of them for modeling.
+
+    Returns
+    -------
+    : numpy array
+        Dataset of the length number_of_neighbors <= length. Each record is created from the position, value and
+        distance to the unknown point ``[[x, y, value, distance to unknown position]]``.
+    """
+    # Distances to unknown point
+    r = np.array([unknown_position])
+
+    known_pos = data_array[:, :-1]
+    dists = calc_point_to_point_distance(r, known_pos)
+    angles = calc_angles(known_pos)
+    angle_diffs = calculate_angular_distance(angles, direction)
+
+    selected_data = select_possible_neighbors_angular(data_array,
+                                                      dists,
+                                                      angle_diffs,
+                                                      neighbors_range,
+                                                      number_of_neighbors,
+                                                      use_all_neighbors_in_range)
+
+    return selected_data[:, :-1]
+
+
 def select_kriging_data(unknown_position: Iterable,
                         data_array: np.ndarray,
                         neighbors_range: float,
@@ -486,18 +611,13 @@ def select_kriging_data(unknown_position: Iterable,
     sorted_neighbors_and_dists = neighbors_and_dists[neighbors_and_dists[:, -1].argsort()]
     prepared_data = sorted_neighbors_and_dists[sorted_neighbors_and_dists[:, -1] <= neighbors_range, :]
 
-    len_prep = len(prepared_data)
-    is_more_than_0 = len_prep > 0
-
-    if use_all_neighbors_in_range and is_more_than_0:
-        return prepared_data
-
-    if len_prep > number_of_neighbors:
-        return sorted_neighbors_and_dists[:number_of_neighbors]
-    elif not is_more_than_0:
-        return sorted_neighbors_and_dists[:number_of_neighbors]
+    if len(prepared_data) >= number_of_neighbors:
+        if use_all_neighbors_in_range:
+            return prepared_data
+        else:
+            return prepared_data[:number_of_neighbors]
     else:
-        return prepared_data
+        return sorted_neighbors_and_dists[:number_of_neighbors]
 
 
 def select_poisson_kriging_data(u_block_centroid: np.ndarray,