+++ pyinterpolate/variogram/regularization/deconvolution.py
@@ -108,8 +108,8 @@ class Deconvolution:
 
         - 0 or 180: is E-W,
         - 90 or 270 is N-S,
-        - 45 or 225 is NW-SE,
-        - 135 or 315 is NE-SW.
+        - 45 or 225 is NE-SW,
+        - 135 or 315 is NW-SE.
 
     tolerance : float (in range [0, 1])
 
@@ -254,7 +254,7 @@ class Deconvolution:
             agg_step_size: float,
             agg_max_range: float,
             agg_nugget: float = 0,
-            agg_direction: float = 0,
+            agg_direction: float = None,
             agg_tolerance: float = 1,
             variogram_weighting_method: str = "closest",
             model_types: Union[str, List] = 'basic') -> None:
@@ -293,8 +293,8 @@ class Deconvolution:
 
             - 0 or 180: is E-W,
             - 90 or 270 is N-S,
-            - 45 or 225 is NW-SE,
-            - 135 or 315 is NE-SW.
+            - 45 or 225 is NE-SW,
+            - 135 or 315 is NW-SE.
 
         agg_tolerance : float (in range [0, 1]), optional, default=1
             If ``agg_tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
@@ -531,7 +531,7 @@ class Deconvolution:
                       agg_step_size: float,
                       agg_max_range: float,
                       agg_nugget: float = 0,
-                      agg_direction: float = 0,
+                      agg_direction: float = None,
                       agg_tolerance: float = 1,
                       variogram_weighting_method: str = "closest",
                       model_types: Union[str, List] = 'basic',
@@ -573,8 +573,8 @@ class Deconvolution:
 
             - 0 or 180: is E-W,
             - 90 or 270 is N-S,
-            - 45 or 225 is NW-SE,
-            - 135 or 315 is NE-SW.
+            - 45 or 225 is NE-SW,
+            - 135 or 315 is NW-SE.
 
         agg_tolerance : float (in range [0, 1]), optional, default=1
             If ``agg_tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of